core: fix slots equivocation tests
